Fixed bad JEXL/elConfig example in README.md
The JEXL/elConfig example in the README.md documentation will not work
to read the database password from a file path pointed to by an
environment variable because the first argument to
org.apache.commons.io.FileUtils.readFileToString() must be an instance
of java.io.File and not String. This commit fixes that with an example
that is tested to work with Grouper 2.4.x.